backend/requirements.txt
openai>=1.0.0
requests>=2.31.0
pandas>=2.0.0
feedparser>=6.0.0
beautifulsoup4>=4.12.0
arxiv>=1.4.0
python-dotenv>=1.0.0
pytest>=7.4.0
